fix (dev_spring_boot) renamed auth classes
